Move commit-buildr functions to commit/builder.rs
